rename webSocket package names
