+++ example/java/example/src/main/java/io/bitmax/api/websocket/messages/requests/CancelOrder.java
@@ -1,4 +1,4 @@
-package io.bitmax.api.websocket.messages.outcome;
+package io.bitmax.api.websocket.messages.requests;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
