+++ example/java/example/src/main/java/io/bitmax/api/websocket/messages/responses/Trade.java
@@ -1,4 +1,4 @@
-package io.bitmax.api.websocket.messages.income;
+package io.bitmax.api.websocket.messages.responses;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
