+++ example/java/example/src/test/java/io/bitmax/api/examples/account/OrderWebSocketExample.java
@@ -3,9 +3,9 @@ package io.bitmax.api.examples.account;
 import io.bitmax.api.JavaAuthClient;
 import io.bitmax.api.rest.RestBitMax;
 import io.bitmax.api.websocket.BitMaxApiWebSocketListener;
-import io.bitmax.api.websocket.messages.outcome.CancelOrder;
-import io.bitmax.api.websocket.messages.outcome.PlaceOrder;
-import io.bitmax.api.websocket.messages.outcome.Subscribe;
+import io.bitmax.api.websocket.messages.requests.CancelOrder;
+import io.bitmax.api.websocket.messages.requests.PlaceOrder;
+import io.bitmax.api.websocket.messages.requests.Subscribe;
 
 import java.util.Map;
